Ajout de la classe AppManager et changement de int à long les attributs "id"

+++ app/src/main/java/fr/jerome/climbinggymlog/database/ClientDB.java
@@ -74,7 +74,7 @@ public class ClientDB extends DBHandler {
      */
     public void delete(Client client) {
 
-        int id = client.getId();
+        long id = client.getId();
         System.out.println("Client supprimé avec l'id: " + id);
         database.delete(TABLE_NAME, ID_U  + " = " + id, null);
